requirements.txt
# Core dependencies
numpy>=1.21.0
PyQt5>=5.15.4
Pillow>=8.3.0

# Deep Learning
torch>=1.9.0
torchvision>=0.10.0

# Computer Vision
opencv-python>=4.5.3
mediapipe>=0.8.9

# Audio Processing
librosa>=0.8.1
sounddevice>=0.4.3

# Machine Learning
scikit-learn>=0.24.2

# Utilities
tqdm>=4.62.0
matplotlib>=3.4.3
pandas>=1.3.0

# Optional CUDA support (uncomment if using GPU)
# torch-cuda>=1.9.0
# torchvision-cuda>=0.10.0
